Adding conditional logic

+++ linux_scripts/ubuntu_server_setup.sh
@@ -20,7 +20,13 @@ echo "
 Step 2: Add gpg Key
 ###################################
 "
-curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+if [ -e /usr/share/keyrings/docker-archive-keyring.gpg]
+then
+	echo "GPG Key previously added. Skipping."
+else
+	curl -fsSL https://download.docker.com/linux/ubuntu/gpg | sudo gpg --dearmor -o /usr/share/keyrings/docker-archive-keyring.gpg
+	echo "GPG Key added."
+fi
 
 echo "deb [arch=$(dpkg --print-architecture) signed-by=/usr/share/keyrings/docker-archive-keyring.gpg] https://download.docker.com/linux/ubuntu $(lsb_release -cs) stable" | sudo tee /etc/apt/sources.list.d/docker.list > /dev/null
 
@@ -37,22 +43,39 @@ Step 4: Install docker-ce
 ###################################
 "
 
-apt-cache policy docker-ce
+if ! command -v docker &> /dev/null
+then
 
-sudo apt install docker-ce -y
+	apt-cache policy docker-ce
+
+	sudo apt install docker-ce -y
+
+	sudo usermod -aG docker ${USER}
+
+	echo "Docker Installed."
+else
+	echo "Docker previously installed. Skipping."
+fi
+
+docker --version
 
 echo "
 ###################################
-Step 5: Set docker user perms
+Step 5: Install docker compose
 ###################################
 "
 
-sudo usermod -aG docker ${USER}
+if ! command -v docker compose &> /dev/null
+then
+	mkdir -p ~/.docker/cli-plugins/
 
-mkdir -p ~/.docker/cli-plugins/
+	curl -SL https://github.com/docker/compose/releases/download/v2.3.3/docker-compose-linux-armv6 -o ~/.docker/cli-plugins/docker-compose
+	
+	chmod +x ~/.docker/cli-plugins/docker-compose
 
-curl -SL https://github.com/docker/compose/releases/download/v2.3.3/docker-compose-linux-armv6 -o ~/.docker/cli-plugins/docker-compose
-
-chmod +x ~/.docker/cli-plugins/docker-compose
+	echo "Docker Compose Installed."
+else
+	echo "Docker Compose previously installed. Skipping."
+fi
 
 docker compose version
